+++ cli/axonopscli/components/dashboard.py
@@ -8,7 +8,7 @@ class Dashboard:
     def __init__(self, axonops, args):
         self.axonops = axonops
         self.args = args
-        self.full_dashboard_url = f"{DASHBOARD_TEMPLATE_URL}/{args.org}/cassandra/{args.cluster}?dashver=2.0"
+        self.full_dashboard_url = f"{DASHBOARD_TEMPLATE_URL}/{args.org}/{args.cluster_type}/{args.cluster}?dashver=2.0"
         self.dashboard_data = None
 
     def get_actual_dashboards(self):
@@ -125,7 +125,7 @@ class Dashboard:
 
             # Update AxonOps with the new list of dashboards
             update_payload = {
-                'type': 'cassandra',
+                'type': self.args.cluster_type,
                 'dashboards': self.dashboard_data
             }
             response = self.axonops.do_request(
@@ -153,7 +153,7 @@ class Dashboard:
 
         # Update AxonOps with the new list of dashboards
         update_payload = {
-            'type': 'cassandra',
+            'type': self.args.cluster_type,
             'dashboards': self.dashboard_data
         }
         response = self.axonops.do_request(